+++ packages/react-graphql-universal-provider/package.json
@@ -28,7 +28,6 @@
     "@shopify/react-graphql": "^5.1.4",
     "@shopify/react-hooks": "^1.2.3",
     "@shopify/react-html": "^9.2.3",
-    "@shopify/react-network": "^3.2.2",
     "apollo-client": "^2.3.8"
   },
   "peerDependencies": {
