+++ packages/react-graphql-universal-provider/src/index.ts
@@ -1,5 +1,4 @@
 export {
-  GRAPHQL_OPERATIONS,
   GraphQLClientOptions,
   GraphQLUniversalProvider,
 } from './GraphQLUniversalProvider';
