+++ packages/react-graphql/src/graphql-operations-context.ts
@@ -1,0 +1,8 @@
+import React from 'react';
+import {GraphQLOperationDetails} from './types';
+
+export const GRAPHQL_OPERATIONS = Symbol('graphQLOperations');
+
+export const GraphQLOperationsContext = React.createContext<
+  GraphQLOperationDetails[] | []
+>([]);
