+++ packages/react-graphql/src/index.ts
@@ -8,9 +8,14 @@ export {
   GraphQLVariables,
   GraphQLDeepPartial,
   QueryProps,
+  GraphQLOperationDetails,
 } from './types';
 
 export {ApolloProvider} from './ApolloProvider';
+export {
+  GraphQLOperationsContext,
+  GRAPHQL_OPERATIONS,
+} from './graphql-operations-context';
 
 export * from './hooks';
 export * from './links';
