+++ packages/react-network/src/hooks.ts
@@ -6,10 +6,6 @@ import {useServerEffect} from '@shopify/react-effect';
 import {NetworkContext} from './context';
 import {NetworkManager} from './manager';
 
-export function useNetworkManager() {
-  return useContext(NetworkContext);
-}
-
 export function useNetworkEffect(perform: (network: NetworkManager) => void) {
   const network = useContext(NetworkContext);
