+++ packages/react-network/src/server.ts
@@ -1,5 +1,5 @@
 import {Context} from 'koa';
-import {NetworkManager, EFFECT_ID, STATE_ID} from './manager';
+import {NetworkManager, EFFECT_ID} from './manager';
 
 export {NetworkContext} from './context';
 export {NetworkManager, EFFECT_ID};
@@ -8,7 +8,7 @@ export function applyToContext<T extends Context>(
   ctx: T,
   manager: NetworkManager,
 ) {
-  const {status, redirectUrl, headers, serverState} = manager.extract();
+  const {status, redirectUrl, headers} = manager.extract();
 
   if (redirectUrl) {
     ctx.redirect(redirectUrl);
@@ -22,11 +22,5 @@ export function applyToContext<T extends Context>(
     ctx.set(header, value);
   }
 
-  ctx.state[STATE_ID] = serverState;
-
   return ctx;
-}
-
-export function getServerState<T extends Context>(ctx: T) {
-  return ctx.state[STATE_ID];
 }
