+++ packages/react-server/package.json
@@ -28,6 +28,7 @@
     "@shopify/react-async": "^3.0.8",
     "@shopify/react-csrf-universal-provider": "^1.0.4",
     "@shopify/react-effect": "^3.2.3",
+    "@shopify/react-graphql": "^5.1.4",
     "@shopify/react-html": "^9.2.3",
     "@shopify/react-hydrate": "^1.1.4",
     "@shopify/react-network": "^3.2.2",
